Add sample data, load it on the map
